% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmod.R
\name{regdiag}
\alias{regdiag}
\title{Title Regression Goodness of Fit}
\usage{
regdiag(x)
}
\arguments{
\item{x}{Numeric Value}
}
\value{
Dataframe of Goodness of Fit
}
\description{
Title Regression Goodness of Fit
}
\examples{
regdia(model1)
}
